Made my theme let the background be transparent.
You see, I've already setup transparency in the terminal thing. So I
just needed the theme to 'allow' it I guess?

ANyways, I didn't have a problem until I was ssh'd into a server using
it that way, at which point the theme of nvim was no longer transparent.
TOggling this value fixes that.